requirements.txt
fastapi
uvicorn
requests
google-api-python-client
google-auth-oauthlib
google-auth-httplib2
transformers
torch
pymongo
dateparser